operating/canva_automation/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
beautifulsoup4==4.12.2
python-pptx==0.6.23
jinja2==3.1.2

# API-specific dependencies (no browser automation needed for these)
# For Canva Connect API
urllib3>=1.26.0
